+++ src/dynamics/prognostic_variables.jl
@@ -268,36 +268,59 @@ function set!(
 end
 
 # set LTA <- LTA 
-function set!(var::LowerTriangularArray{T}, L::LowerTriangularArray, varargs...; add::Bool) where T
+function set!(
+    var::LowerTriangularArray,
+    L::LowerTriangularArray,
+    varargs...;
+    add::Bool=false,
+)
     if add 
         if size(var) == size(L)
-            var .+= T.(L) 
+            var .+= L
         else 
             L_var = spectral_truncation(L, size(var, 1, as=Matrix), size(var, 2, as=Matrix))
             var .+= L_var
         end 
     else 
-        size(var) != size(L) || fill!(var, zero(T)) # copyto! copies over the largest subset, when size(var) > size(L), the copyto! isn't enough by itself
+        size(var) != size(L) || fill!(var, 0) # copyto! copies over the largest subset, when size(var) > size(L), the copyto! isn't enough by itself
         copyto!(var, L)
     end 
     return var
 end 
 
 # set LTA <- Grid 
-function set!(var::LowerTriangularArray, grids::AbstractGridArray, geometry::Union{Geometry, Nothing}=nothing, S::Union{Nothing, SpectralTransform}=nothing; add)
+function set!(
+    var::LowerTriangularArray,
+    grids::AbstractGridArray,
+    geometry::Union{Geometry, Nothing}=nothing,
+    S::Union{Nothing, SpectralTransform}=nothing;
+    add::Bool=false,
+)
     specs = isnothing(S) ? transform(grids) : transform(grids, S)
     set!(var, specs; add)
 end
 
 # set LTA <- func 
-function set!(var::LowerTriangularArray, f::Function, geometry::Geometry{NF}, S::Union{SpectralTransform, Nothing}=nothing; add::Bool) where NF
-    grid = ndims(var) == 1 ? zeros(geometry.Grid{NF}, geometry.nlat_half) : zeros(geometry.Grid{NF}, geometry.nlat_half, geometry.nlayers)
+function set!(
+    var::LowerTriangularArray,
+    f::Function,
+    geometry::Geometry{NF, Grid},
+    S::Union{SpectralTransform, Nothing}=nothing;
+    add::Bool=false,
+) where {NF, Grid}
+    grid = ndims(var) == 1 ? zeros(Grid{NF}, geometry.nlat_half) : zeros(Grid{NF}, geometry.nlat_half, geometry.nlayers)
     set!(grid, f, geometry, S; add=false)
     set!(var, grid, geometry, S; add)
 end
 
 # set LTA <- number
-function set!(var::LowerTriangularArray{T}, s::Number, geometry::Geometry{NF}, S::Union{SpectralTransform, Nothing}=nothing; add::Bool) where {T, NF}
+function set!(
+    var::LowerTriangularArray{T},
+    s::Number,
+    geometry::Geometry{NF},
+    S::Union{SpectralTransform, Nothing}=nothing;
+    add::Bool=false,
+) where {T, NF}
     
     # appropiate normalization, assume standard 2√π normalisation if no transform is given 
     norm_sphere = isnothing(S) ? 2sqrt(π) : S.norm_sphere
@@ -313,7 +336,13 @@ function set!(var::LowerTriangularArray{T}, s::Number, geometry::Geometry{NF}, S
 end 
 
 # set Grid <- Grid
-function set!(var::AbstractGridArray, grids::AbstractGridArray, geometry::Geometry, S::Union{Nothing, SpectralTransform}=nothing; add)
+function set!(
+    var::AbstractGridArray,
+    grids::AbstractGridArray,
+    geometry::Geometry,
+    S::Union{Nothing, SpectralTransform}=nothing;
+    add::Bool=false,
+)
     if add 
         if grids_match(var, grids)
             var .+= grids
@@ -327,13 +356,25 @@ function set!(var::AbstractGridArray, grids::AbstractGridArray, geometry::Geomet
 end 
 
 # set Grid <- LTA
-function set!(var::AbstractGridArray, specs::LowerTriangularArray, geometry::Geometry, S::Union{Nothing, SpectralTransform}=nothing; add)
+function set!(
+    var::AbstractGridArray,
+    specs::LowerTriangularArray,
+    geometry::Geometry,
+    S::Union{Nothing, SpectralTransform}=nothing;
+    add::Bool=false,
+)
     grids = isnothing(S) ? transform(specs) : transform(specs, S)
     set!(var, grids, geometry, S; add)
 end
 
 # set Grid <- Func
-function set!(var::AbstractGridArray, f::Function, geometry::Geometry, S::Union{Nothing, SpectralTransform}=nothing; add)
+function set!(
+    var::AbstractGridArray,
+    f::Function,
+    geometry::Geometry,
+    S::Union{Nothing, SpectralTransform}=nothing;
+    add::Bool=false,
+)
     (; londs, latds, σ_levels_full) = geometry
     kernel(a, b) = add ? a+b : b
     for k in eachgrid(var)
@@ -350,7 +391,7 @@ function set!(
     f::Function,
     geometry::Geometry,
     S::Union{Nothing, SpectralTransform}=nothing;
-    add
+    add::Bool=false,
 ) where T
     (; londs, latds) = geometry
     kernel(a, b) = add ? a+b : b
@@ -366,7 +407,7 @@ function set!(
     s::Number, 
     geometry::Union{Geometry, Nothing}=nothing, 
     S::Union{Nothing, SpectralTransform}=nothing;
-    add::Bool,
+    add::Bool=false,
 ) where T
     kernel(a, b) = add ? a+b : b
     sT = T(s)
@@ -381,7 +422,7 @@ function set_vordiv!(
     v_func,
     geometry::Geometry,
     S::Union{Nothing, SpectralTransform}=nothing;
-    add::Bool,
+    add::Bool=false,
     coslat_scaling_included::Bool=false,
 )
     u_L = similar(vor) 
@@ -400,7 +441,7 @@ function set_vordiv!(
     v::AbstractGridArray,
     geometry::Geometry,
     S::Union{Nothing, SpectralTransform}=nothing;
-    add::Bool,
+    add::Bool=false,
     coslat_scaling_included::Bool=false,
 )
     u_ = coslat_scaling_included ? u : RingGrids.scale_coslat⁻¹(u)
@@ -420,7 +461,7 @@ function set_vordiv!(
     v::LowerTriangularArray,
     geometry::Geometry,
     S::Union{Nothing, SpectralTransform}=nothing;
-    add::Bool,
+    add::Bool=false,
     coslat_scaling_included::Bool=false,
 ) 
     S = isnothing(S) ? SpectralTransform(geometry.spectral_grid) : S